preparem for reuse of getEmployees in graophql resolver
